Allow to pass listeners as instances to `robot.run` and `TestSuite.run`.
