+++ utest/running/test_running.py
@@ -8,6 +8,7 @@ from os.path import abspath, dirname, normpath, join
 from robot.utils.asserts import assert_equals
 from robot.running import TestSuite, TestSuiteBuilder
 from resources.runningtestcase import RunningTestCase
+from resources.Listener import Listener
 
 
 CURDIR = dirname(abspath(__file__))
@@ -253,6 +254,12 @@ class TestStateBetweenTestRuns(unittest.TestCase):
 
 
 class TestListeners(RunningTestCase):
+
+    def test_listeners(self):
+        module_file = join(ROOTDIR, 'utest', 'resources', 'Listener.py')
+        suite = build('setups_and_teardowns.robot')
+        suite.run(output=None, log=None, report=None, listener=[module_file+":1", Listener(2)])
+        self._assert_outputs([("[from listener 1]", 1), ("[from listener 2]", 1)])
 
     def test_listeners_unregistration(self):
         module_file = join(ROOTDIR, 'utest', 'resources', 'Listener.py')
