+++ utest/utils/test_text.py
@@ -1,10 +1,12 @@
 import unittest
+import os
+from os.path import abspath
 
 from robot.utils.asserts import *
 
 from robot.utils.text import cut_long_message, _count_line_lengths, \
     _MAX_ERROR_LINES, _MAX_ERROR_LINE_LENGTH, _ERROR_CUT_EXPLN,\
-    get_console_length, pad_console_length, split_tags_from_doc
+    get_console_length, pad_console_length, split_tags_from_doc, split_args_from_name_or_path
 
 
 class NoCutting(unittest.TestCase):
@@ -195,6 +197,77 @@ class TestDocSplitter(unittest.TestCase):
             self._assert_doc_and_tags(original, exp_doc, exp_tags)
             self._assert_doc_and_tags(original+'\n', exp_doc, exp_tags)
 
+
+class TestSplitArgsFromNameOrPath(unittest.TestCase):
+
+    def setUp(self):
+        self.method = split_args_from_name_or_path
+
+    def test_with_no_args(self):
+        assert not os.path.exists('name'), 'does not work if you have name folder!'
+        assert_equals(self.method('name'), ('name', []))
+
+    def test_with_args(self):
+        assert not os.path.exists('name'), 'does not work if you have name folder!'
+        assert_equals(self.method('name:arg'), ('name', ['arg']))
+        assert_equals(self.method('listener:v1:v2:v3'), ('listener', ['v1', 'v2', 'v3']))
+        assert_equals(self.method('aa:bb:cc'), ('aa', ['bb', 'cc']))
+
+    def test_empty_args(self):
+        assert not os.path.exists('foo'), 'does not work if you have foo folder!'
+        assert_equals(self.method('foo:'), ('foo', ['']))
+        assert_equals(self.method('bar:arg1::arg3'), ('bar', ['arg1', '', 'arg3']))
+        assert_equals(self.method('3:'), ('3', ['']))
+
+    def test_semicolon_as_separator(self):
+        assert_equals(self.method('name;arg'), ('name', ['arg']))
+        assert_equals(self.method('name;1;2;3'), ('name', ['1', '2', '3']))
+        assert_equals(self.method('name;'), ('name', ['']))
+
+    def test_alternative_separator_in_value(self):
+        assert_equals(self.method('name;v:1;v:2'), ('name', ['v:1', 'v:2']))
+        assert_equals(self.method('name:v;1:v;2'), ('name', ['v;1', 'v;2']))
+
+    def test_windows_path_without_args(self):
+        assert_equals(self.method('C:\\name.py'), ('C:\\name.py', []))
+        assert_equals(self.method('X:\\APPS\\listener'), ('X:\\APPS\\listener', []))
+        assert_equals(self.method('C:/varz.py'), ('C:/varz.py', []))
+
+    def test_windows_path_with_args(self):
+        assert_equals(self.method('C:\\name.py:arg1'), ('C:\\name.py', ['arg1']))
+        assert_equals(self.method('D:\\APPS\\listener:v1:b2:z3'),
+                      ('D:\\APPS\\listener', ['v1', 'b2', 'z3']))
+        assert_equals(self.method('C:/varz.py:arg'), ('C:/varz.py', ['arg']))
+        assert_equals(self.method('C:\\file.py:arg;with;alternative;separator'),
+                      ('C:\\file.py', ['arg;with;alternative;separator']))
+
+    def test_windows_path_with_semicolon_separator(self):
+        assert_equals(self.method('C:\\name.py;arg1'), ('C:\\name.py', ['arg1']))
+        assert_equals(self.method('D:\\APPS\\listener;v1;b2;z3'),
+                      ('D:\\APPS\\listener', ['v1', 'b2', 'z3']))
+        assert_equals(self.method('C:/varz.py;arg'), ('C:/varz.py', ['arg']))
+        assert_equals(self.method('C:\\file.py;arg:with:alternative:separator'),
+                      ('C:\\file.py', ['arg:with:alternative:separator']))
+
+    def test_existing_paths_are_made_absolute(self):
+        path = 'robot-framework-unit-test-file-12q3405909qasf'
+        open(path, 'w').close()
+        try:
+            assert_equals(self.method(path), (abspath(path), []))
+            assert_equals(self.method(path+':arg'), (abspath(path), ['arg']))
+        finally:
+            os.remove(path)
+
+    def test_existing_path_with_colons(self):
+        # Colons aren't allowed in Windows paths (other than in "c:")
+        if os.sep == '\\':
+            return
+        path = 'robot:framework:test:1:2:42'
+        os.mkdir(path)
+        try:
+            assert_equals(self.method(path), (abspath(path), []))
+        finally:
+            os.rmdir(path)
 
 if __name__ == '__main__':
     unittest.main()
